feat: complete sosnina_a_diff_count task with MPI and sequential implementations

+++ tasks/sosnina_a_diff_count/mpi/src/ops_mpi.cpp
@@ -59,7 +59,7 @@ bool SosninaADiffCountMPI::RunImpl() {
   std::size_t start =
       (static_cast<std::size_t>(rank) * block_size) + std::min(static_cast<std::size_t>(rank), remainder);
   std::size_t end = start + block_size;
-  if (static_cast<std::size_t>(rank) < remainder) {
+  if (rank < static_cast<int>(remainder)) {
     end += 1;
   }
   end = std::min(end, total_len);
